requirements.txt
python-binance==1.0.16
pandas==1.5.3
numpy==1.24.3
ccxt==3.0.74
ta==0.10.2
pydantic==1.10.8
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
pytest==7.3.1
schedule==1.2.0
tqdm==4.65.0 